size_t type corrections within the scope of the autoexpirer
This standardizes several uses of size_t and long long to int64_t, for
the purposes of resolving a casting issue that occurs in the
autoexpiration code on 32-bit systems.

This bumps the library ABI version.

+++ mythtv/libs/libmythbase/filesysteminfo.cpp
@@ -58,7 +58,7 @@ FileSystemInfo::FileSystemInfo(const FileSystemInfo &other)
 }
 
 FileSystemInfo::FileSystemInfo(QString hostname, QString path, bool local,
-        int fsid, int groupid, int blksize, long long total, long long used) :
+        int fsid, int groupid, int blksize, int64_t total, int64_t used) :
     m_hostname(hostname), m_path(path), m_local(local), m_fsid(fsid),
     m_grpid(groupid), m_blksize(blksize), m_total(total), m_used(used)
 {
@@ -172,7 +172,7 @@ const QList<FileSystemInfo> FileSystemInfo::RemoteGetInfo(MythSocket *sock)
 }
 
 void FileSystemInfo::Consolidate(QList<FileSystemInfo> &disks,
-                                 bool merge, size_t fuzz)
+                                 bool merge, int64_t fuzz)
 {
     int newid = 0;
 
@@ -194,15 +194,15 @@ void FileSystemInfo::Consolidate(QList<FileSystemInfo> &disks,
 
             int bSize = max(32, max(it1->getBlockSize(), it2->getBlockSize())
                                         / 1024);
-            long long diffSize = it1->getTotalSpace() - it2->getTotalSpace();
-            long long diffUsed = it1->getUsedSpace() - it2->getUsedSpace();
+            int64_t diffSize = it1->getTotalSpace() - it2->getTotalSpace();
+            int64_t diffUsed = it1->getUsedSpace() - it2->getUsedSpace();
 
             if (diffSize < 0)
                 diffSize = 0 - diffSize;
             if (diffUsed < 0)
                 diffUsed = 0 - diffUsed;
 
-            if ((diffSize <= bSize) && ((size_t)diffUsed <= fuzz))
+            if ((diffSize <= bSize) && (diffUsed <= fuzz))
             {
                 it2->setFSysID(it1->getFSysID());
 
@@ -224,7 +224,7 @@ void FileSystemInfo::Consolidate(QList<FileSystemInfo> &disks,
 
 void FileSystemInfo::PopulateDiskSpace(void)
 {
-    long long total = -1, used = -1;
+    int64_t total = -1, used = -1;
     getDiskSpace(getPath().toAscii().constData(), total, used);
     setTotalSpace(total);
     setUsedSpace(used);
