+++ mythtv/libs/libmythbase/storagegroup.cpp
@@ -655,10 +655,10 @@ QString StorageGroup::FindFileDir(QString filename)
 QString StorageGroup::FindNextDirMostFree(void)
 {
     QString nextDir;
-    long long nextDirFree = 0;
-    long long thisDirTotal;
-    long long thisDirUsed;
-    long long thisDirFree;
+    int64_t nextDirFree = 0;
+    int64_t thisDirTotal;
+    int64_t thisDirUsed;
+    int64_t thisDirFree;
 
     LOG(VB_FILE, LOG_DEBUG, LOC + QString("FindNextDirMostFree: Starting"));
